--- a/Source/DafnyCore/Verifier/Datatypes/BoogieGenerator.DataTypes.cs
+++ b/Source/DafnyCore/Verifier/Datatypes/BoogieGenerator.DataTypes.cs
@@ -503,8 +503,9 @@ namespace Microsoft.Dafny {
               */
             Bpl.Expr ct = FunctionCall(ctor.Origin, ctor.FullName, Predef.DatatypeType, args);
             var rhs = FunctionCall(ctor.Origin, BuiltinFunction.DtRank, null, ct);
-            var trigger = BplTrigger(rhs);
-            q = new Bpl.ForallExpr(ctor.Origin, bvs, trigger, Bpl.Expr.Lt(lhs, rhs));
+            var trigger = BplTrigger(ct);
+            var kv = new QKeyValue(ctor.Origin, "weight", [ Expr.Literal(3) ]);
+            q = new Bpl.ForallExpr(ctor.Origin, [], bvs, kv, trigger, Bpl.Expr.Lt(lhs, rhs));
             AddOtherDefinition(fn, new Bpl.Axiom(ctor.Origin, q, "Inductive rank"));
           } else if (argType is SeqType) {
             // axiom (forall params, i: int {#dt.ctor(params)} :: 0 <= i && i < |arg| ==> DtRank(arg[i]) < DtRank(#dt.ctor(params)));