… Source/DafnyPipeline.Test/expectedProverLog.smt2